+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2
+
+### Added
+- Encrypt/Decrypt sensitive data stored in shared preferences
+- Add or update multiple key-value pairs in a single batch operation
+- Add or remove listeners for shared preference changes
+
 ## 0.0.1+1
 
 ### Added